# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dota Data API",
    "version": "1.0.0",
    "description": "API documentation for Dota Data service routes."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "OpenDotaHero": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "localized_name": {
            "type": "string"
          },
          "primary_attr": {
            "type": "string"
          },
          "attack_type": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "img": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "base_health": {
            "type": "integer"
          },
          "base_mana": {
            "type": "integer"
          },
          "base_armor": {
            "type": "number"
          },
          "base_attack_min": {
            "type": "integer"
          },
          "base_attack_max": {
            "type": "integer"
          },
          "move_speed": {
            "type": "integer"
          },
          "base_attack_time": {
            "type": "number"
          },
          "attack_point": {
            "type": "number"
          },
          "attack_range": {
            "type": "integer"
          },
          "projectile_speed": {
            "type": "integer"
          },
          "turn_rate": {
            "type": "number"
          },
          "cm_enabled": {
            "type": "boolean"
          },
          "legs": {
            "type": "integer"
          },
          "day_vision": {
            "type": "integer"
          },
          "night_vision": {
            "type": "integer"
          },
          "hero_id": {
            "type": "integer"
          },
          "turbo_picks": {
            "type": "integer"
          },
          "turbo_wins": {
            "type": "integer"
          },
          "pro_ban": {
            "type": "integer"
          },
          "pro_win": {
            "type": "integer"
          },
          "pro_pick": {
            "type": "integer"
          },
          "1_pick": {
            "type": "integer"
          },
          "1_win": {
            "type": "integer"
          },
          "2_pick": {
            "type": "integer"
          },
          "2_win": {
            "type": "integer"
          },
          "3_pick": {
            "type": "integer"
          },
          "3_win": {
            "type": "integer"
          },
          "4_pick": {
            "type": "integer"
          },
          "4_win": {
            "type": "integer"
          },
          "5_pick": {
            "type": "integer"
          },
          "5_win": {
            "type": "integer"
          },
          "6_pick": {
            "type": "integer"
          },
          "6_win": {
            "type": "integer"
          },
          "7_pick": {
            "type": "integer"
          },
          "7_win": {
            "type": "integer"
          },
          "8_pick": {
            "type": "integer"
          },
          "8_win": {
            "type": "integer"
          },
          "null_pick": {
            "type": "integer"
          },
          "null_win": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/heroes": {
      "get": {
        "tags": [
          "Heroes"
        ],
        "summary": "Get all heroes",
        "responses": {
          "200": {
            "description": "List of heroes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenDotaHero"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Hero data not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Heroes"
        ],
        "summary": "Fetch and return all heroes (synchronous)",
        "responses": {
          "200": {
            "description": "List of heroes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenDotaHero"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}